fix: Remove transfomers
