+++ package.json
@@ -47,6 +47,7 @@
     "typescript": "^4.5.2"
   },
   "dependencies": {
+    "@dcl/schemas": "^3.0.0",
     "axios": "^0.24.0",
     "dcl-catalyst-commons": "^7.1.3",
     "dcl-crypto": "^2.3.0",
